removing outdated, duplicate code in device/configuration and device/adapter (#37)
