+++ src/chariot/device/configuration/ImpinjXArrayConfiguration.py
@@ -1,43 +1,34 @@
 from typing import Dict, List, Type
-
 from chariot.JSONTypes import JSONDict, JSONObject
 from chariot.device.configuration.DeviceConfiguration import DeviceConfiguration
 
 
 class ImpinjXArrayConfiguration(DeviceConfiguration):
+    requiredFields: Dict[str, Type[JSONObject]] = {
+        'ipAddress': str,
+        'authUsername': str,
+        'authPassword': str,
+    }
+    optionalFields: Dict[str, Type[JSONObject]] = {
+        'jobName': str,
+        'facility': str,
+        'startDelay': str,
+        'readerGroups': List[str],
+        'reportToDatabaseEnabled': bool,
+        'reportToHistoryEnabled': bool,
+        'reportToMessageQueueEnabled': bool,
+        'useOtherJobData': bool,
+    }
+
     def __init__(self, configMap: JSONDict):
+        self.requiredFields.update(super().requiredFields)
+        self.optionalFields.update(super().optionalFields)
         super().__init__(configMap)
 
-    def _validateConfig(self, configMap: JSONDict) -> None:
-        super()._validateConfig(configMap)
+    def _validateInitialConfig(self, configMap: JSONDict) -> None:
+        super()._validateInitialConfig(configMap)
         if configMap['deviceType'] != 'Impinj xArray':
             raise AssertionError
-
-        requiredFields: Dict[str, Type[JSONObject]] = {
-            'ipAddress': str,
-            'authUsername': str,
-            'authPassword': str,
-        }
-        optionalFields: Dict[str, Type[JSONObject]] = {
-            'jobName': str,
-            'facility': str,
-            'startDelay': str,
-            'readerGroups': List[str],
-            'reportToDatabaseEnabled': bool,
-            'reportToHistoryEnabled': bool,
-            'reportToMessageQueueEnabled': bool,
-            'useOtherJobData': bool,
-        }
-
-        for field, fieldType in requiredFields:
-            if field not in configMap or not isinstance(
-                    configMap[field], fieldType):
-                raise ValueError
-
-        for field, fieldType in optionalFields:
-            if field in configMap and not isinstance(
-                    configMap[field], fieldType):
-                raise ValueError
 
 
 __all__ = ['ImpinjXArrayConfiguration']
